Rebase and fix typo
